fix: url-http-create-request complains about multibytes characters
When the xml argument of the function xml-rpc-request contains
multibytes characters, the value stored in the variable
url-request-data is not properly encoded as a unibyte character
string.

The subsequent call to url-http-create-request fail with the error
message "Multibyte text in HTTP request: [...]", in the package url-http.

This bug occures with emacs 25.2 under a fresh debian 9.3 install,
and all packages from ELPA.

It's probably the same bug which occures here:
https://github.com/proofit404/anaconda-mode/issues/189 .

This fix seams to fix the issue.